feat(all): ✨ restructuring folders
