+++ src/resources/config/url-prod.yml
@@ -1,0 +1,1 @@
+url: "https://valentinos-magic-beans.click"
