+++ src/resources/config/url-qa.yml
@@ -1,0 +1,1 @@
+url: "https://valentinos-magic-beans.click"
