+++ src/resources/data/prod/credencial.yml
@@ -1,0 +1,3 @@
+valid_user:
+  email: "scrooge-mcduck@uorak.com"
+  password: "MoneyBin$1886"
